Refactor to separate cache service.
